\name{GFDrmtlGUI}
\alias{GFDrmtlGUI}
\title{
A Shiny app for multiple RMTL-based tests
}
\description{
This function provides a shiny app for performing multiple RMTL-based tests. Here, the asymptotic or permutation tests can be applied.
}
\usage{
GFDrmtlGUI()
}
\details{
Note that the calculations for the permutation test may take a while, especially with stepwise extension.
}
\value{
No return value, called for side effects.
}
